20220317 提交未提交代码

+++ biz/mitutucute-springboot-controller/src/main/java/com/xiaozhuanglt/mitutucute/springboot/controller/AmapServiceController.java
@@ -22,9 +22,9 @@ public class AmapServiceController {
      * @Param: [areaId]
      * @return {@link int}
      */
-    public int queryArea(Long areaId){
+    public Long queryArea(Long areaId){
 
-        int i = serviceDemo.queryArea(areaId);
+        Long i = serviceDemo.queryArea(areaId);
         return i;
     }
